Move tally_metric feature to just metrics and minor TS improvements
