+++ features/features.go
@@ -3,7 +3,7 @@ package features
 import (
 	"go.temporal.io/sdk-features/features/activity/retry_on_error"
 	"go.temporal.io/sdk-features/features/bugs/go/child_workflow_cancel_panic"
-	"go.temporal.io/sdk-features/features/telemetry/tally_metrics"
+	"go.temporal.io/sdk-features/features/telemetry/metrics"
 	"go.temporal.io/sdk-features/harness/go/harness"
 )
 
@@ -11,6 +11,6 @@ func init() {
 	harness.MustRegisterFeatures(
 		child_workflow_cancel_panic.Feature,
 		retry_on_error.Feature,
-		tally_metrics.Feature,
+		metrics.Feature,
 	)
 }
