+++ features/telemetry/metrics/README.md
@@ -1,5 +1,5 @@
-# Tally metrics
+# Metrics
 
-In Temporal, the Tally library can be used in some SDKs to record metrics.
+In Temporal, the SDKs record metrics and provide metric interfaces to record your own.
 
 This feature workflow runs two activities and adds custom counters. Then the custom counters are checked at the end.
